migracie fixed

+++ AN-API/database/migrations/2023_10_16_150901_create_devices_in_networks_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('devices_in_networks', function (Blueprint $table) {
             $table->id('device_id');
             $table->string('name');
-            $table->unsignedInteger('router_id');
+            $table->unsignedBigInteger('router_id');
             $table->timestamps();
         });
